// src/interfaces/noticeIntefaces.ts
export interface INotice{ 
    _id: string;
    avatar: string;
    category: string;
    title: string;
    name: string;
    birthDate: string;
    breed: string;
    sex: string;
    location: string;
    comments: string;
    price: string;
    owner: string;
    email: string;
    phoneNumber: string;
}

export interface ICreatedNotice {
    avatar: string;
    category: string;
    title: string;
    name: string;
    birthDate: string;
    breed: string;
    sex: string;
    location: string;
    comments: string;
    price: string;
    owner: string;
    email: string;
    phoneNumber: string;
}